scaie_crm/core/backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
sqlalchemy>=1.4.0,<2.0.0
python-dotenv>=0.19.0
openai>=1.0.0
pydantic>=2.0.0
python-telegram-bot>=20.0
requests>=2.25.0
httpx>=0.25.0
python-multipart>=0.0.5